fix(reader): stop truncating /ByteRange offsets to int
PdfSignature.ByteRange was int[], filled from PdfInteger.Value — a long — through an unchecked
(int) cast. /ByteRange holds byte offsets into the file, so for a signed PDF larger than 2 GB
every value past int.MaxValue wrapped silently: 3,000,000,000 became -1,294,967,296 and
4,294,967,296 became 0. Nothing reported an error, so the signature was checked against the
wrong bytes and the ISO 19005-2 §6.4.3-1 coverage rule evaluated nonsense.

The same truncating parse existed twice: once in PdfDocumentReader for AcroForm signatures and
again in SignatureRule.ExtractSignature for the catalog /Perms /DocMDP path. Two independent
parsers for one field is how issue #167's bug came to exist in two places at once, so both are
fixed here.

ByteRange is now ReadOnlyMemory<long>, matching Contents in the same type. Read-only because the
property handed out the internal array unaliased, so a caller could mutate the reader's state.
Reader is still Preview and its surface still sits in PublicAPI.Unshipped.txt, so this costs
nothing — which is the argument for doing it now rather than freezing the broken shape later.

The regression test builds a small PDF whose signature dictionary claims offsets past
int.MaxValue rather than a multi-gigabyte fixture: the defect was in parsing the array, not in
reading a large file, so the parse is what needs exercising. It asserts the absence of the
specific wrapped values, so it documents the failure mode and not just the fix. Mutation-checked:
restoring the (int) cast fails it.

PdfDocumentReader already had CheckedOffset for exactly this hazard on xref offsets, whose
comment notes such a field "can hold a value larger than int.MaxValue" and that it throws
"rather than wrapping silently". /ByteRange was the one place that practice was not applied.

Five copies of the same Buffer.BlockCopy pair across four test files are replaced by one
ReconstructSignedContent helper, which keeps the narrowing to int that Buffer.BlockCopy needs in
a single checked place — so a fixture over 2 GB would throw there rather than quietly copy the
wrong bytes.

Verified: Kernel 895, Layout 217, Reader 185, Cli 344, and Conformance 1,236 with veraPDF
available and nothing skipped, which covers the SignatureRule change against the oracle.

+++ src/VellumPdf.Reader/PdfDocumentReader.cs
@@ -542,14 +542,17 @@ public sealed class PdfDocumentReader : IDisposable
             subFilter = sfName;
 
         var brObj = sigDict.Get(new PdfName("ByteRange"));
-        int[] byteRange = [];
+        long[] byteRange = [];
         if (brObj is PdfArray brArr)
         {
-            byteRange = new int[brArr.Count];
+            byteRange = new long[brArr.Count];
             for (var i = 0; i < brArr.Count; i++)
             {
+                // PdfInteger.Value is already a long, and these are file offsets: narrowing to int
+                // silently wrapped every offset past 2 GB, so a large signed file was checked
+                // against the wrong byte ranges without any error being reported.
                 if (brArr[i] is PdfInteger pi)
-                    byteRange[i] = (int)pi.Value;
+                    byteRange[i] = pi.Value;
             }
         }
